perf(fields/x86): use bit-spread for Gf2_128 square — 2× faster than CLMUL
PCLMUL-based squaring pays four dependent 7-cycle CLMULs (2 for the
square, 2 for modular reduction), while the bit-spread approach is a
chain of single-cycle shift/mask ops with excellent ILP. Measured on
this workstation:

  gf2_128/square  (CLMUL sq128 + CLMUL reduce):  11.5 ns
  gf2_128/square  (bit-spread + scalar reduce):   6.0 ns

Reuses the same algorithm the soft backend already uses — in char 2,
`(Σ aᵢ xⁱ)² = Σ aᵢ x^(2i)`, so squaring is just "spread each bit to
twice its index" with no multiplies at all.

The `square_xmm` helper in the inverse loop stays on CLMUL: state is
already in XMM there, and leaving XMM per iteration to do scalar
bit-spread would cost more than it saves.

Also moves `bit_spread_u32` out of `bmul.rs` into a new always-compiled
`spread` module, since it's now needed by both the soft *and* x86
backends. `bmul` remains gated to the soft path (carry-less multiply
primitives are only used when no hardware clmul is available).
